Create Code-Challenge-Database-Creation.py
Approach:
So I have used MySQL Database for data ingestion purpose, Amongst the options of PostgreSQL and other databases MySQL is better for memory optimization. Since there wasn't much of a large number of data size hence proceeded further with MySQL.

Enhancements: Given a choice of usage of cloud service especially AWS we can store the wx_data nd yx_data in S3 and then load it to Redshift for better scalability and performance. As the time required for loading records from S3 to any other component is quite minimal.